requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
email-validator>=2.1.0
httpx>=0.26.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
